Added word user mapping repository

+++ Discord.WordGraffiti/DAL/Models/UserWordMapping.cs
@@ -3,7 +3,7 @@
     public class UserWordMapping
     {
         public int Id { get; set; }
-        public int UserId { get; set; }
+        public ulong UserId { get; set; }
         public int WordId { get; set; }
         public int Value { get; set; }
     }
